+++ includes/integrations/gutenberg.php
@@ -19,6 +19,7 @@ add_action('init', function() {
             'show_view_toggle' => ['type' => 'string', 'default' => 'true'],
             'show_scale_toggle' => ['type' => 'string', 'default' => 'true'],
             'show_periods' => ['type' => 'string', 'default' => 'true'],
+            'show_data_warnings' => ['type' => 'string', 'default' => 'true'],
             'country_order' => ['type' => 'string', 'default' => ''],
             'colors'   => ['type' => 'string', 'default' => '#336DFF, #AF9BE2, #3B82F6, #BEE00F, #7FD3F7']
         ]